GUAC-1170: Move modules back to separate packages.
